[AArch64][SVE] Add SubtargetFeature to disable lowering unpredicated loads/stores as LDR/STR
PR #127837 changed the lowering for unpredicated loads/stores to use LDR/STR instead of LD1/ST1.
However, on some CPUs, such as A64FX, there is a performance difference between LD1/ST1 and LDR/STR.
As a result, the lowering introduced in #127837 can cause a performance regression on these targets.
This patch adds a SubtargetFeature to disable this lowering and prevent the regression.